[#111]; refactor(mail): storage레이어 저장로직 단순화
